change warning text
The warning text is now changed to primarily deal with the missing
configure_requires prerequisites.

Mention the possibility that the installation of a configure_requires
module failed.

Mention the upgrade installations only for CPAN.pm and CPANPLUS users;
it's unlikely that there are cpanminus versions which cannot deal with
configure_requires (probably it's also unlikely for CPANPLUS, but I
did not check this).

Remove the paragraph about running Build.PL instead; if the user is
really doing a manual install, then running Build.PL will not help
anything, the prerequisites will still be missing. Quite the contrary
it's likely that the user will just get a less-informative
perl-generated error because of a missing module.

Remove the authorship of the warning text.

Remove the "Congralutations" text for the case that there are no
configure_requires problems; there's no need to mention this.