do nothing during "build", since Build.PL will do that for us
This was never caught before because we did not usually print the
warning during build, unless there was a config prereq missing.  Now
that we always print the warning when Makefile.PL is run, this needless
double build is more obvious.

'dzil test' operations are not affected (even the double testing with
RELEASE_TESTING on and off, one for each build) because ->test calls
MakeMaker::Runner->test which calls MakeMaker::Runner->build, rather
than calling our build.